refactor report filter/result page UI for adjust spacing/padding
