+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/report/ReportFilterPage.kt
@@ -17,11 +17,15 @@
 package org.smartregister.fhircore.anc.ui.report
 
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material.Button
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -59,9 +63,18 @@ fun ReportFilterPage(
   Surface(color = colorResource(id = R.color.white)) {
     Column(modifier = Modifier.fillMaxSize()) {
       TopBarBox(topBarTitle, onBackPress)
-      DateSelectionBox(startDate, endDate, true, onStartDatePress, onEndDatePress)
-      PatientSelectionBox(patientSelectionText, selectedPatient, onPatientSelectionTypeChanged)
-      GenerateReportButton(generateReportEnabled, onGenerateReportPress)
+      Box(modifier = Modifier.padding(16.dp)) {
+        Column {
+          DateSelectionBox(startDate, endDate, true, onStartDatePress, onEndDatePress)
+          Spacer(modifier = Modifier.size(16.dp))
+          PatientSelectionBox(patientSelectionText, selectedPatient, onPatientSelectionTypeChanged)
+          Column(modifier = Modifier.fillMaxHeight(), verticalArrangement = Arrangement.Bottom) {
+            Row(modifier = Modifier.fillMaxWidth(), verticalAlignment = Alignment.Bottom) {
+              GenerateReportButton(generateReportEnabled, onGenerateReportPress)
+            }
+          }
+        }
+      }
     }
   }
 }
@@ -94,6 +107,42 @@ fun ReportFilterScreen(viewModel: ReportViewModel) {
 @Composable
 @Preview(showBackground = true)
 @ExcludeFromJacocoGeneratedReport
+fun PreviewDateRangeChangable() {
+  DateSelectionBox(startDate = "Start date", endDate = "End date", canChange = true)
+}
+
+@Composable
+@Preview(showBackground = true)
+@ExcludeFromJacocoGeneratedReport
+fun PreviewDateRangeFixed() {
+  DateSelectionBox(startDate = "Start date", endDate = "End date", canChange = false)
+}
+
+@Composable
+@Preview(showBackground = true)
+@ExcludeFromJacocoGeneratedReport
+fun PreviewPatientSelectionAll() {
+  PatientSelectionBox(
+    patientSelectionText = ReportViewModel.PatientSelectionType.ALL,
+    onPatientSelectionChange = {},
+    selectedPatient = PatientItem()
+  )
+}
+
+@Composable
+@Preview(showBackground = true)
+@ExcludeFromJacocoGeneratedReport
+fun PreviewPatientSelectionIndividual() {
+  PatientSelectionBox(
+    patientSelectionText = ReportViewModel.PatientSelectionType.INDIVIDUAL,
+    onPatientSelectionChange = {},
+    selectedPatient = PatientItem(name = "Ind Patient Item")
+  )
+}
+
+@Composable
+@Preview(showBackground = true)
+@ExcludeFromJacocoGeneratedReport
 fun ReportFilterPreview() {
   ReportFilterPage(
     topBarTitle = "PageTitle",
@@ -112,24 +161,17 @@ fun ReportFilterPreview() {
 
 @Composable
 fun GenerateReportButton(generateReportEnabled: Boolean, onGenerateReportClicked: () -> Unit) {
-  Row(
-    horizontalArrangement = Arrangement.SpaceBetween,
-    modifier = Modifier.fillMaxWidth().padding(vertical = 20.dp),
-    verticalAlignment = Alignment.Bottom
-  ) {
-    Column(modifier = Modifier.align(Alignment.Bottom)) {
-      Button(
-        enabled = generateReportEnabled,
-        onClick = onGenerateReportClicked,
-        modifier =
-          Modifier.fillMaxWidth().padding(horizontal = 16.dp).testTag(REPORT_GENERATE_BUTTON)
-      ) {
-        Text(
-          color = Color.White,
-          text = stringResource(id = R.string.generate_report),
-          modifier = Modifier.padding(8.dp)
-        )
-      }
+  Column {
+    Button(
+      enabled = generateReportEnabled,
+      onClick = onGenerateReportClicked,
+      modifier = Modifier.fillMaxWidth().padding(horizontal = 16.dp).testTag(REPORT_GENERATE_BUTTON)
+    ) {
+      Text(
+        color = Color.White,
+        text = stringResource(id = R.string.generate_report),
+        modifier = Modifier.padding(8.dp)
+      )
     }
   }
 }
